Clippy friend

+++ server/src/game/world/generation.rs
@@ -1,7 +1,7 @@
 use bevy::prelude::*;
 use common::game::world::chunk::*;
-use noise::{BasicMulti, HybridMulti, MultiFractal, NoiseFn, OpenSimplex, RidgedMulti};
-use rand::Rng;
+use noise::{MultiFractal, NoiseFn, OpenSimplex, RidgedMulti};
+
 
 pub fn generate_chunk(pos: IVec3, seed: u32) -> RawChunk {
     //TODO: Switch to using ron files to determine biomes and what blocks they should use. For now hardcoding a simplex noise
@@ -19,18 +19,18 @@ pub fn generate_chunk(pos: IVec3, seed: u32) -> RawChunk {
                 if full_y as f64 <= noise_val && full_y as f64 >= (noise_val - 2.0) {
                     raw_chunk.add_block_state(&"vinoxgrass".to_string());
                     raw_chunk.set_block(
-                        UVec3::new(x as u32, y as u32, z as u32),
+                        UVec3::new(x, y, z),
                         "vinoxgrass".to_string(),
                     );
                 } else if full_y as f64 <= noise_val {
                     raw_chunk.add_block_state(&"vinoxdirt".to_string());
                     raw_chunk.set_block(
-                        UVec3::new(x as u32, y as u32, z as u32),
+                        UVec3::new(x, y, z),
                         "vinoxdirt".to_string(),
                     );
                 } else {
                     raw_chunk
-                        .set_block(UVec3::new(x as u32, y as u32, z as u32), "air".to_string());
+                        .set_block(UVec3::new(x, y, z), "air".to_string());
                 }
 
                 // let multi_noise = ridged_noise.get([
